refactor: standardize naming to jac
