+++ docarray/store/__init__.py
@@ -1,5 +1,5 @@
 from docarray.store.file import FileDocStore
-from docarray.store.jinaai import JACDocStore
+from docarray.store.jac import JACDocStore
 from docarray.store.s3 import S3DocStore
 
 __all__ = ['JACDocStore', 'FileDocStore', 'S3DocStore']
