+++ docarray/store/jac.py
